Move final functionality from Presenter class and remove the Presenter class.
